change app router on contact
